authentication/requirements.txt
asgiref==3.6.0
certifi
Django==3.2.17
django-cors-headers==3.5.0
djangorestframework
djangorestframework-simplejwt==4.4.0
djongo==1.3.6
dnspython==2.3.0
mysqlclient==2.0.1
PyJWT==1.7.1
pymongo==3.12.0
PyMySQL==1.0.2
pytz
sqlparse==0.2.4
typing_extensions==4.4.0
wincertstore==0.2
